buyer/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
pyjwt>=2.8.0
pydantic>=2.6.0
python-dotenv>=1.0.0
chromadb>=0.5.0
sentence-transformers>=2.7.0
